Add Marketly customization termination comment
To follow the established conventions in the data file, this adds
specific comments which mark the beginning and end of the section of
TLDs that have been moved specifically for the customization of
behavior by Marketly.

This also drops a few temporary comments and shifts a small additional
set of entries into the Marketly override section.